Persistence and UI changes.
Use prepared statements to enforce type-safe queries.
Save Temperature values to persistence.

+++ jweatherhistory-model/src/main/java/za/co/johanmynhardt/jweatherhistory/model/RainEntry.java
@@ -4,7 +4,7 @@ package za.co.johanmynhardt.jweatherhistory.model;
  * @author Johan Mynhardt
  */
 public class RainEntry {
-	public final int id;
+	public final long id;
 	public final int volume;
 	public final String description;
 	private WeatherEntry weatherEntry;
